+++ mllam_data_prep/create_dataset.py
@@ -4,6 +4,7 @@ from collections import defaultdict
 from pathlib import Path
 
 import numpy as np
+import pyproj
 import xarray as xr
 import zarr
 from loguru import logger
@@ -15,6 +16,13 @@ from .ops.chunking import chunk_dataset
 from .ops.derive_variable import derive_variable
 from .ops.loading import load_input_dataset
 from .ops.mapping import map_dims_and_variables
+from .ops.projection import (
+    ProjectionCompatibilityWarning,
+    ProjectionInconsistencyWarning,
+    check_projection_compatibility,
+    get_projection_crs,
+    validate_projection_consistency,
+)
 from .ops.selection import select_by_kwargs
 from .ops.statistics import calc_stats
 from .ops.subsetting import extract_variable
@@ -137,6 +145,7 @@ def create_dataset(config: Config):
     chunking_config = config.output.chunking
 
     dataarrays_by_target = defaultdict(list)
+    projections = {}
 
     for dataset_name, input_config in config.inputs.items():
         path = input_config.path
@@ -145,6 +154,7 @@ def create_dataset(config: Config):
         target_output_var = input_config.target_output_variable
         expected_input_attributes = input_config.attributes
         expected_input_var_dims = input_config.dims
+        dataset_projections = input_config.projections or {}
 
         output_dims = output_config.variables[target_output_var]
 
@@ -192,6 +202,40 @@ def create_dataset(config: Config):
             dataset_name=dataset_name,
         )
 
+        # get the projection information from the dataset and update it with the projection
+        # information given in the input config
+        # TODO: read the projection information from the config
+        try:
+            projection_crs = get_projection_crs(ds)
+        except ValueError:
+            projection_crs = None
+        if projection_crs is None and dataset_projections:
+            logger.warning(
+                f"Projection information not found in dataset {dataset_name}, using projection information from config."
+            )
+            projection_crs = {
+                crs_var: {"crs_wkt": p.crs_wkt}
+                for crs_var, p in dataset_projections.items()
+            }
+        elif projection_crs is None and not dataset_projections:
+            logger.warning(
+                f"No projection information found neither in dataset {dataset_name}, nor in config."
+            )
+        elif projection_crs is not None and dataset_projections:
+            logger.warning(
+                f"Projection information given in {dataset_name} is overwritten by that given in config."
+            )
+            projection_crs = {
+                crs_var: {"crs_wkt": p.crs_wkt}
+                for crs_var, p in dataset_projections.items()
+            }
+        if projection_crs is not None:
+            projections.update(projection_crs)
+
+        # TODO: remove this once grid_mapping is set in config file
+        for var in ds.data_vars:
+            ds[var].attrs["grid_mapping"] = list(projection_crs.keys())[0]
+
         dim_mapping = input_config.dim_mapping
 
         # check that there is an entry for each arch dimension
@@ -232,6 +276,24 @@ def create_dataset(config: Config):
             da_target = select_by_kwargs(da_target, **selection_kwargs)
 
         dataarrays_by_target[target_output_var].append(da_target)
+
+    # validate projections across input datasets
+    if projections:
+        try:
+            validate_projection_consistency(projections.values())
+        except ProjectionInconsistencyWarning as e:
+            logger.warning(f"Projection information might be ambiguous: {e}")
+        projection = pyproj.CRS.from_cf(list(projections.values())[0])
+        try:
+            check_projection_compatibility(list(projections.values())[0])
+        except ProjectionCompatibilityWarning:
+            logger.warning(
+                "WKT string should include a BBOX definition to be compatible with e.g. cartopy."
+            )
+
+        # TODO: generalize the retrieval of x and y coords
+        # coords = (dataarrays_by_target[target_output_var][0]['x'], dataarrays_by_target[target_output_var][0]['y'])
+        # lon, lat = get_latitude_longitude_from_projection(projection, coords)
 
     ds = _merge_dataarrays_by_target(dataarrays_by_target=dataarrays_by_target)
 
@@ -277,6 +339,14 @@ def create_dataset(config: Config):
             coords={"split_name": list(splits.keys()), "split_part": ["start", "end"]},
         )
         ds["splits"] = da_splits
+
+    # add the projection information to the dataset
+    if projections:
+        ds[projections.keys()] = xr.DataArray(0, attrs=projection.to_cf()).astype(
+            "int16"
+        )
+        # for var in ["VARIABLES_WITH_PROKJECTION"]:
+        #     ds[var].attrs["grid_mapping"] = "__common__"
 
     ds.attrs = {}
     ds.attrs["schema_version"] = config.schema_version